Fix calculations in the new observation page
When searching for all the passes of a satellite above a station we
were checking if the next observation's elevation is above the defined
elevation in station's settings. In case that wasn't, loop for the
next observation was broken.

In the case above, if a low pass was first, next passes inside the
defined time span never been returned and never shown as observations.
With this commit, when elevation is less than the defined in station
settings, then calculations continue with the next pass of the
satellite above the examined station.