Add overlap check after schedule button is hit
When schedule button was hit, calculated observations were scheduled
without any check if between calculation and scheduling has been any
change on the already scheduled station observations.

This resulted to scheduling of the same observation more than one
time, if you opened and calculated the same observation in two
different tabs and then hit schedule button.

With this commit, this issue if fixed as a check is done when the
schedule button is hit. If there is an overlap it returns an error
message which inform the user that there are one or more overlaps.