Fix behavior of the new observation page if overlaps exists
The refactoring changed the behavior of the new observation page if overlaps exists:
Previously, if any overlap occured an error was returned and not any observation was scheduled.
This changed so it returns an error but schedules all the observations until it hits the error.

This commit restores the previous behavior.