Fix calculation of pass details
When creating the observation with a given a start and end time, the functionality of pyephems next_pass isn't required (and could result in pass details for a pass not corresponding to the given start and end time). Instead, the pass details can be calculated directly for the given start, mid-pass and end time.

Fixes #479.