Fix disappearing mark on the map in station view
Sometimes mark on the map wasn't visible for the initial zoom. This was
result of a race condition between loading the pin image (loadImage and
addImage) and adding the layer with the mark on the map (addLayer).

Adding addLayer call inside the loadImage callback solves the race
condition by having the pin image always loaded before adding the layer.

Signed-off-by: Alfredos-Panagiotis Damkalis <fredy@fredy.gr>