Change annotated queries for better performance
Annotaded queries in observations and stations pages on observations table
had as a result to have a "LEFT OUTER JOIN" with all the fields of
"SELECT" being in the "GROUP BY" which creates a significant delay on
running the query.

In django documentation it is stated that "values()" should be used to
specify the fields in "GROUP BY" but we needed other fields and properties
from the Station and User models. With this commit we change the annotate
to alternative queries that solve the significant delay.

Signed-off-by: Alfredos-Panagiotis Damkalis <fredy@fredy.gr>