tests/ObservationFactory: Choose from existing satellites & tles
The ObservationFactory called a subfactory "SatelliteFactory" to
generate satellites with random norad ids (in the 2000 - 4000 range).
This range of NORAD IDs can be found in public catalogs, but most of
those satellites are already decayed (and thus they are not part of
any Celestrak list).

Running update_all_tle thus creates many requests to Celestrak
(~<= 200, the number of observation fixtures) by trying to fetch their
TLEs via the satcat/tle.php endpoint.

This commit solves that problem by adding a RealisticObservationFactory
which uses only existing Satellites (fetched from satnogs-db).

A small caveat: Assigned Satellite & TLE for a generated observation
most likely wont match (but this was already the case before).

Another small caveat: The RealisticObservationFactory can generate at
max only observations as there are satellite (a limitation of the
underlying factory.Iterator).

Requires the TLEs to be fetched before the fixtures can be generated.

Fixes #695.

Signed-off-by: Fabian P. Schmidt <kerel@mailbox.org>