base/scheduling: Make division operations "true division"-safe
Use explicit typecast to float to prevent ambiguous division operation
in the first two cases. In the last case (timdelta, int), the python2
timedelta doesn't support the division operator (this was intruduced
in python 3.2). Use old_div instead.

This change clarifies that true division is the desired behaviour
in these three code locations.

Signed-off-by: Fabian P. Schmidt <kerel@mailbox.org>