Fix waterfall status from kept vetted status
On migration of adding waterfall status, initialization of data from
vetted status was done falsely on observations that have no waterfall
and observations with waterfall weren't initialized with any value.

In this commit we fix this issue by initializing from vetted status
the waterfall status on observations that have waterfall and we move
waterfall status back to "unknown" (with value "None") on all the
observations that don't have waterfall artifact.

Signed-off-by: Alfredos-Panagiotis Damkalis <fredy@fredy.gr>